Поправил задание №1 из 10 лекции, вынес шифровщик/дешифровщик в отдельный модуль, алфавит в отдельный файл
