+++ Practice/S.Mikheev/Lec_10/alphabet.txt
@@ -1,0 +1,1 @@
+ abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ.,:-!
